+++ CHANGELOG.md
@@ -116,3 +116,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Only host (first joined player) can start the game (messagetype)
 - Readme
 
+- new network package for test
+- - MockSocket